Refactor Baro to floats, filter at position rate
convert pressure to altitude early
remove median filter
PT2 filtering upsampled to altitude function in position.c - thanks KarateBrot
baro task synced to position task - thanks Steve
PT2 implementation - thanks KarateBrot
ground altitude from filtered altitude
baro cali by average of calibration samples over cal period
adjust vario and smoothing defaults
don't say haveBaroAlt until cal is complete
reduce PIDs since Baro is faster
add baro smoothing values to blackbox header

+++ src/test/unit/telemetry_ibus_unittest.cc
@@ -82,7 +82,7 @@ int32_t getAmperage(void)
     return amperage;
 }
 
-int32_t getEstimatedVario(void)
+int16_t getEstimatedVario(void)
 {
     return estimatedVario;
 }
